Fix some shadow declarations and CodeQL warning
Possible fix for
https://github.com/netpanzer/netpanzer/security/code-scanning/30 related
to #104

For naming, I'm going by the Google style guide at
https://google.github.io/styleguide/cppguide.html#Variable_Names which
is mentioned in our contributing document.

+++ src/Lib/2D/Surface.hpp
@@ -58,8 +58,8 @@ class Surface : public NoCopy {
   void setOffsetCenter();
 
   // Accessor functions.
-  unsigned int getWidth() const { return twidth; }
-  unsigned int getHeight() const { return theight; }
+  unsigned short getWidth() const { return twidth; }
+  unsigned short getHeight() const { return theight; }
   unsigned int getPitch() const { return tpitch; }
 
   iXY getCenter() const { return iXY(getWidth() >> 1, getHeight() >> 1); }
